Allow leading space in syslog message
By default rsyslog maintains a leading space in the received syslog
message to comply with RFC3164. The current setup on front uses the
mmrm1stspace rsyslog module to strip this space, but that's not
available in the rsyslog on CentOS 7. Be flexible in the parser to
support either rsyslog configuration.