Handle non-delta downloads
Now that we store the root treedir id for each commit in the cache
we can also handle non-delta downloads, as long as each toplevel treedir
is unique (which is pretty likely, or at least that they are the same ref).

For every log line that is a treedir download we do a reverse map
lookup to find the commit. This only works if the commit is already in
the cache, which could happen by accident due to a delta download, but
we also prime the cache with the current commit for the specified ref,
which is likely to work ok for "fresh" log files.